Squashed commit

+++ keycloak/openid_client.go
@@ -26,7 +26,7 @@ type OpenidClientSecret struct {
 type OpenidClientAuthorizationSettings struct {
 	PolicyEnforcementMode         string `json:"policyEnforcementMode,omitempty"`
 	DecisionStrategy              string `json:"decisionStrategy,omitempty"`
-	AllowRemoteResourceManagement bool   `json:"allowRemoteResourceManagement,omitempty"`
+	AllowRemoteResourceManagement bool   `json:"allowRemoteResourceManagement"`
 	KeepDefaults                  bool   `json:"-"`
 }
 
@@ -63,6 +63,7 @@ type OpenidClient struct {
 
 type OpenidClientAttributes struct {
 	PkceCodeChallengeMethod                  string                           `json:"pkce.code.challenge.method"`
+	RequireDPoPBoundTokens                   types.KeycloakBoolQuoted         `json:"dpop.bound.access.tokens,omitempty"`
 	ExcludeSessionStateFromAuthResponse      types.KeycloakBoolQuoted         `json:"exclude.session.state.from.auth.response"`
 	ExcludeIssuerFromAuthResponse            types.KeycloakBoolQuoted         `json:"exclude.issuer.from.auth.response"`
 	AccessTokenLifespan                      string                           `json:"access.token.lifespan"`
@@ -151,9 +152,13 @@ func (keycloakClient *KeycloakClient) NewOpenidClient(ctx context.Context, clien
 			if err != nil {
 				return err
 			}
-			err = keycloakClient.DeleteOpenidClientAuthorizationResource(ctx, resource.RealmId, resource.ResourceServerId, resource.Id)
-			if err != nil {
-				return err
+			// Only attempt to delete if the default resource exists
+			// (Keycloak 26.5+ doesn't create default resources automatically)
+			if resource != nil {
+				err = keycloakClient.DeleteOpenidClientAuthorizationResource(ctx, resource.RealmId, resource.ResourceServerId, resource.Id)
+				if err != nil {
+					return err
+				}
 			}
 		}
 	}
@@ -203,6 +208,15 @@ func (keycloakClient *KeycloakClient) GetOpenidClient(ctx context.Context, realm
 
 	client.RealmId = realmId
 	client.ClientSecret = clientSecret.Value
+
+	if client.AuthorizationServicesEnabled {
+		var authSettings OpenidClientAuthorizationSettings
+		err = keycloakClient.get(ctx, fmt.Sprintf("/realms/%s/clients/%s/authz/resource-server", realmId, id), &authSettings, nil)
+		if err != nil {
+			return nil, err
+		}
+		client.AuthorizationSettings = &authSettings
+	}
 
 	return &client, nil
 }
